# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transaction Service API",
        "description": "API de gestion des transactions",
        "contact": {
            "email": "n.taffot@elyft.tech"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8003",
            "description": "Transaction Service Server"
        }
    ],
    "paths": {
        "/api/transactions/commission-settings": {
            "get": {
                "tags": [
                    "Commission Settings"
                ],
                "summary": "Lister toutes les commissions",
                "operationId": "ef52f328bff09ed6877d972e7f251a73",
                "responses": {
                    "200": {
                        "description": "Liste des commissions"
                    },
                    "400": {
                        "description": "Requête invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Requête invalide"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Non authentifié"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Accès refusé"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 403
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "PERMISSION_DENIED"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ressource non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ressource non trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Commission Settings"
                ],
                "summary": "Créer une nouvelle règle de commission",
                "operationId": "a3fc262a8eb947ec60e9484f39757194",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "entreprise_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "country_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "operator_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "transaction_type": {
                                        "type": "string",
                                        "enum": [
                                            "deposit",
                                            "withdrawal"
                                        ]
                                    },
                                    "commission_type": {
                                        "type": "string",
                                        "enum": [
                                            "percentage",
                                            "fixed"
                                        ]
                                    },
                                    "commission_value": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "min_amount": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "max_amount": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "created_by": {
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé avec succès"
                    },
                    "400": {
                        "description": "Requête invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Requête invalide"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Non authentifié"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Accès refusé"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 403
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "PERMISSION_DENIED"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ressource non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ressource non trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}